backend_python/requirements.txt
flask
mysql-connector-python
python-dotenv
flask-cors
bcrypt
